Fix column type

+++ plugins/tracker/db/install.sql
@@ -194,7 +194,7 @@ CREATE TABLE tracker_changeset_comment(
     submitted_by INT(11) NULL,
     submitted_on INT(11) NOT NULL,
     body TEXT NOT NULL,
-    body_format INT(11) NOT NULL DEFAULT '0',
+    body_format TINYINT(1) NOT NULL DEFAULT '0',
     old_artifact_history_id INT(11) NULL,
     INDEX changeset_idx(changeset_id)
 );
@@ -591,4 +591,3 @@ INSERT INTO permissions_values (permission_type,ugroup_id) VALUES ('PLUGIN_TRACK
 INSERT INTO permissions_values (permission_type,ugroup_id) VALUES ('PLUGIN_TRACKER_WORKFLOW_TRANSITION',2);
 INSERT INTO permissions_values (permission_type,ugroup_id,is_default) VALUES ('PLUGIN_TRACKER_WORKFLOW_TRANSITION',3,1);
 INSERT INTO permissions_values (permission_type,ugroup_id) VALUES ('PLUGIN_TRACKER_WORKFLOW_TRANSITION',4);
-
